Perhitungan Baru Sampe Di Keputusan Awal

+++ database/migrations/2023_01_18_150603_create_subkriterias_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('kriteria_id')->constrained()->onDelete('cascade');
             $table->string('subkriteria');
-            $table->float('nilai');
+            $table->int('nilai');
             $table->timestamps();
         });
     }
